commonize WebWorkerDriver. rename `createWebWorkerDriver()` to `createDefaultWebWorkerDriver()`. remove unnecessary `jsCommon` group from web-worker-driver.
