+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/CreateWebWorkerDriver.kt
@@ -1,0 +1,5 @@
+package app.cash.sqldelight.driver.worker
+
+import app.cash.sqldelight.db.SqlDriver
+
+expect fun createDefaultWebWorkerDriver(): SqlDriver
