+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/WebWorkerException.kt
