+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/api/WorkerResult.kt
@@ -1,0 +1,3 @@
+package app.cash.sqldelight.driver.worker.api
+
+internal expect interface WorkerResult
