+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/api/WorkerResultWithRowCount.kt
@@ -1,0 +1,6 @@
+package app.cash.sqldelight.driver.worker.api
+
+internal interface WorkerResultWithRowCount {
+  val result: WorkerResult
+  val rowCount: Long
+}
