+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/expected/CheckWorkerResults.kt
@@ -1,0 +1,5 @@
+package app.cash.sqldelight.driver.worker.expected
+
+import app.cash.sqldelight.driver.worker.api.WorkerResult
+
+internal expect fun checkWorkerResults(results: WorkerResult?): WorkerResult
