+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/expected/JsWorkerSqlCursor.kt
@@ -1,0 +1,6 @@
+package app.cash.sqldelight.driver.worker.expected
+
+import app.cash.sqldelight.db.SqlCursor
+import app.cash.sqldelight.driver.worker.api.WorkerResult
+
+internal expect class JsWorkerSqlCursor(result: WorkerResult) : SqlCursor
