+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/expected/JsWorkerSqlPreparedStatement.kt
@@ -1,0 +1,5 @@
+package app.cash.sqldelight.driver.worker.expected
+
+import app.cash.sqldelight.db.SqlPreparedStatement
+
+internal expect class JsWorkerSqlPreparedStatement() : SqlPreparedStatement
