+++ drivers/web-worker-driver/src/commonMain/kotlin/app/cash/sqldelight/driver/worker/expected/Worker.kt
@@ -1,0 +1,3 @@
+package app.cash.sqldelight.driver.worker.expected
+
+expect class Worker
