--- drivers/web-worker-driver/src/jsCommonMain/kotlin/app/cash/sqldelight/driver/worker/CreateWebWorkerDriver.kt
@@ -1,5 +1,0 @@
-package app.cash.sqldelight.driver.worker
-
-import app.cash.sqldelight.db.SqlDriver
-
-expect fun createWebWorkerDriver(): SqlDriver
