--- drivers/web-worker-driver/src/jsMain/kotlin/app/cash/sqldelight/driver/worker/CreateWebWorkerDriver.kt
@@ -1,8 +1,0 @@
-package app.cash.sqldelight.driver.worker
-
-import app.cash.sqldelight.db.SqlDriver
-import org.w3c.dom.Worker
-
-actual fun createWebWorkerDriver(): SqlDriver {
-  return WebWorkerDriver(Worker(js("""new URL("@cashapp/sqldelight-sqljs-worker/sqljs.worker.js", import.meta.url)""")))
-}
