+++ drivers/web-worker-driver/src/jsMain/kotlin/app/cash/sqldelight/driver/worker/api/WorkerAction.kt
@@ -2,5 +2,5 @@ package app.cash.sqldelight.driver.worker.api
 
 internal actual sealed interface WorkerAction
 
-@Suppress("NOTHING_TO_INLINE", "FunctionName")
+@Suppress("NOTHING_TO_INLINE")
 internal actual inline fun WorkerAction(value: String) = value.unsafeCast<WorkerAction>()
