+++ drivers/web-worker-driver/src/jsMain/kotlin/app/cash/sqldelight/driver/worker/expected/Worker.kt
@@ -1,0 +1,3 @@
+package app.cash.sqldelight.driver.worker.expected
+
+actual typealias Worker = org.w3c.dom.Worker
